Abstract all different charts into their own stimulus controllers
