+++ app/views/podcast_metrics/_agent_card.html.erb
@@ -1,18 +1,16 @@
 <%= turbo_frame_tag turbo_tag.to_s do %>
   <% if agents.present? %>
     <div class="card shadow border-0"
-      data-controller="apex"
-      data-apex-id-value="<%= SecureRandom.uuid %>"
-      data-apex-chart-type-value="bar"
-      data-apex-series-type-value="agents"
-      data-apex-series-data-value="<%= parse_agent_data(agents).to_json %>">
+      data-controller="apex-agents"
+      data-apex-agents-id-value="<%= SecureRandom.uuid %>"
+      data-apex-agents-series-data-value="<%= parse_agent_data(agents).to_json %>">
       <div class="card-header card-header-primary-light d-flex">
         <h5><%= agents_title %></h5>
       </div>
       <div class="card-body">
         <div>
           <div class="col">
-            <div data-apex-target="chart"></div>
+            <div data-apex-agents-target="chart"></div>
           </div>
           <div class="col align-items-center">
             <div class="col my-4">
