+++ app/views/podcast_metrics/_dropdays_card.html.erb
@@ -1,13 +1,10 @@
 <%= turbo_frame_tag "dropdays" do %>
   <div class="row d-flex flex-wrap gx-4"
-    data-controller="apex"
-    data-apex-id-value="<%= SecureRandom.uuid %>"
-    data-apex-series-type-value="dropdays"
-    data-apex-interval-value=""
-    data-apex-series-data-value="<%= episode_dropdays.to_json %>"
-    data-apex-date-range-value="">
+    data-controller="apex-dropdays"
+    data-apex-dropdays-id-value="<%= SecureRandom.uuid %>"
+    data-apex-dropdays-series-data-value="<%= episode_dropdays.to_json %>">
     <div class="col">
-      <div data-apex-target="chart" data-chart="dropdays"></div>
+      <div data-apex-dropdays-target="chart" data-chart="dropdays"></div>
     </div>
     <div class="col align-items-center">
       <div class="col my-4">
@@ -27,7 +24,7 @@
               <% title = ep[:ep][:title] %>
               <tr class="align-items-center">
                 <td class="d-flex align-items-center">
-                  <input id="<%= title %>" type="checkbox" class="form-check-input mt-0 me-2" checked="true" data-action="click->apex#toggleSeries" data-series="<%= title %>" data-apex-target="episodebox">
+                  <input id="<%= title %>" type="checkbox" class="form-check-input mt-0 me-2" checked="true" data-action="click->apex-dropdays#toggleSeries" data-series="<%= title %>">
                   <span class="material-icons me-2" aria-label="hidden" style="color: <%= line_chart_colors[i] %>">circle</span>
                   <span class="text-nowrap overflow-hidden"><%= title %></span>
                 </td>
