+++ app/views/podcast_metrics/show.html.erb
@@ -2,7 +2,7 @@
   <% content_for :title, "#{@podcast.title} - #{t(".title")}" %>
 
   <%= form_with(url: podcast_metrics_path(@podcast), method: :get, id: "metrics", data: {controller: "click", click_debounce_value: 200}) do |form| %>
-    <div data-controller="apex" data-apex-id-value="<%= SecureRandom.uuid %>">
+    <div data-controller="apex">
       <div class="container col-12 d-flex align-items-center justify-content-between my-4">
         <h1 class="fw-bold my-0 mb-0 me-4">Metrics at a Glance</h1>
         <div class="">
@@ -28,17 +28,7 @@
         </div>
       </div>
       <div class="col mb-4">
-        <div class="row gx-2">
-          <div class="col-4 mb-2">
-            <%# render "agent_card", agents: @agent_apps, agents_title: "App or Browser", turbo_tag: "agent_apps" %>
-          </div>
-          <div class="col-4 mb-2">
-            <%# render "agent_card", agents: @agent_types, agents_title: "Agent Types", turbo_tag: "agent_types" %>
-          </div>
-          <div class="col-4 mb-2">
-            <%# render "agent_card", agents: @agent_os, agents_title: "Operating System", turbo_tag: "agent_os" %>
-          </div>
-        </div>
+        <%= turbo_frame_tag "agents", src: agents_podcast_metrics_path(podcast: @podcast), loading: "lazy" %>
       </div>
     </div>
     <%= form.submit class: "d-none", data: {click_target: "submit"} %>
